requirements.txt
streamlit
pandas
numpy
plotly
Pillow
torch
torchvision
grad-cam
opencv-python-headless
